pyproject.toml
[tool.poetry]
name = "book_and_drook"
version = "0.1.0"
description = ""
authors = ["l4blee <l4blee@yandex.ru>"]

[tool.poetry.dependencies]
python = "3.9.6"
sanic = "20.12.6"
PyJWT = "^2.6.0"
pymongo = {extras = ["srv"], version = "^4.3.3"}
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
